python/python_api_service/requirements.txt
paho-mqtt==1.6.1
Flask==2.2.2
Flask-Cors==3.0.10